+++ Core/Logic/Builders/TxtBuilder.cs
@@ -6,38 +6,36 @@ using Core.Extensions;
 using Core.Types.Book;
 using Microsoft.Extensions.Logging;
 
-namespace Core.Logic.Builders; 
+namespace Core.Logic.Builders;
 
-public class TxtBuilder(Options options, ILogger logger) : BuilderBase(options, logger) {
+public class TxtBuilder(Options options, ILogger logger) : BuilderBase(options, logger)
+{
     protected override string Extension => "txt";
 
-    protected override async Task BuildInternal(Book book, string fileName) {
-        if (File.Exists(fileName)) {
-            File.Delete(fileName);
-        }
-        
+    protected override async Task BuildInternal(Book book, string fileName)
+    {
+        if (File.Exists(fileName)) File.Delete(fileName);
+
         await using var file = File.CreateText(fileName);
-        
-        foreach (var chapter in book.Chapters.Where(c => c.IsValid)) {
+
+        foreach (var chapter in book.Chapters.Where(c => c.IsValid))
+        {
             await file.WriteLineAsync("   " + chapter.Title);
             await file.WriteLineAsync();
 
             var prettyText = chapter.Content.PrettyHtml().AsHtmlDoc().DocumentNode.GetText();
-            using var sr = new StringReader(prettyText); 
-            
-            while (true) {
-                var line = await sr.ReadLineAsync();
-                if (line == null) {
-                    break;
-                }
+            using var sr = new StringReader(prettyText);
 
-                if (string.IsNullOrWhiteSpace(line)) {
-                    continue;
-                }
-                
+            while (true)
+            {
+                var line = await sr.ReadLineAsync();
+                if (line == null) break;
+
+                if (string.IsNullOrWhiteSpace(line)) continue;
+
                 await file.WriteLineAsync("   " + line.Trim());
             }
-            
+
             await file.WriteLineAsync();
         }
     }
